+++ nym-vpn-core/crates/nym-vpn-lib/src/lib.rs
@@ -19,7 +19,7 @@ pub mod service;
 #[cfg(not(target_os = "ios"))]
 pub(crate) mod socks5_proxy;
 mod tunnel_health;
-#[cfg(any(target_os = "ios", target_os = "android"))]
+#[cfg(any(target_os = "ios", target_os = "android", test))]
 pub mod tunnel_provider;
 pub mod tunnel_state_machine;
 mod wg_config;