+++ nym-vpn-core/crates/nym-vpn-lib/src/tunnel_state_machine/tunnel_monitor.rs
@@ -1241,6 +1241,7 @@ impl TunnelMonitor {
                 interface_addresses,
                 remote_addresses: vec![assigned_addresses.entry_mixnet_gateway_ip],
                 mtu,
+                exclude_vpn_app: false,
             };
 
             self.create_tun_device(packet_tunnel_settings).await?
@@ -1941,6 +1942,7 @@ impl TunnelMonitor {
             interface_addresses,
             remote_addresses: vec![entry_endpoint],
             mtu,
+            exclude_vpn_app: false,
         };
 
         let tun_device = self.create_tun_device(packet_tunnel_settings).await?;